+++ src/components/Video/Controls/Controls.module.css
@@ -72,11 +72,3 @@
     }
   }
 }
-
-.shareComponentWrapper{
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-}
