+++ src/components/Video/ShareComponent/ShareComponent.module.scss
@@ -1,0 +1,8 @@
+.shareComponentWrapper{
+  width: 100%;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+
+}
+
